# Conflicts:
#	DEPS
#	gn/toolchain/BUILD.gn
#	include/core/SkFontMgr.h
#	modules/svg/include/SkSVGContainer.h
#	modules/svg/include/SkSVGFeMerge.h
#	modules/svg/include/SkSVGNode.h
#	modules/svg/include/SkSVGPattern.h
#	modules/svg/include/SkSVGPoly.h
#	modules/svg/include/SkSVGRect.h
#	modules/svg/include/SkSVGShape.h
#	modules/svg/include/SkSVGTransformableNode.h
#	modules/svg/include/SkSVGUse.h
#	modules/svg/src/SkSVGCircle.cpp
#	modules/svg/src/SkSVGDOM.cpp
#	modules/svg/src/SkSVGEllipse.cpp
#	modules/svg/src/SkSVGFeMerge.cpp
#	modules/svg/src/SkSVGGradient.cpp
#	modules/svg/src/SkSVGLine.cpp
#	modules/svg/src/SkSVGLinearGradient.cpp
#	modules/svg/src/SkSVGPattern.cpp
#	modules/svg/src/SkSVGRadialGradient.cpp
#	modules/svg/src/SkSVGRect.cpp
#	modules/svg/src/SkSVGRenderContext.cpp
#	modules/svg/src/SkSVGShape.cpp
#	modules/svg/src/SkSVGText.cpp
#	modules/svg/src/SkSVGUse.cpp
#	modules/svg/svg.gni
#	src/codec/SkAvifCodec.cpp
#	src/codec/SkJpegxlCodec.cpp
#	src/core/SkFontMgr.cpp
#	src/ports/SkFontMgr_win_dw.cpp

+++ tests/SkTBlockListTest.cpp
@@ -6,9 +6,9 @@
  */
 
 #include "include/core/SkTypes.h"
-#include "include/private/base/SkAlign.h"
-#include "src/base/SkBlockAllocator.h"
-#include "src/base/SkTBlockList.h"
+#include "include/private/SkAlign.h"
+#include "src/core/SkBlockAllocator.h"
+#include "src/core/SkTBlockList.h"
 #include "tests/Test.h"
 
 #include <cstddef>
@@ -322,7 +322,8 @@ void run_large_increment_test(skiatest::Reporter* reporter) {
     largeIncrement.reserve(10);
     size_t scratchSize = TBlockListTestAccess::ScratchBlockSize(largeIncrement);
     // SkBlockAllocator aligns large blocks to 4k
-    size_t expected = SkAlignTo(kIncrementMax * TBlockListTestAccess::kAddressAlign, (1 << 12));
+    size_t expected = SkAlignTo(kIncrementMax * TBlockListTestAccess::kAddressAlign,
+                                (size_t) (1 << 12));
     REPORTER_ASSERT(reporter, scratchSize == expected);
 }
